% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_message}
\alias{chat_message}
\alias{chat_update}
\alias{chat_delete}
\alias{chat_update.default}
\alias{chat_delete.default}
\alias{chat_update.slackpost}
\alias{chat_delete.slackpost}
\title{Send/Edit/Delete post from a channel}
\usage{
chat_message(channel, ..., token = Sys.getenv("SLACK_API_TOKEN"))

chat_update(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)

chat_delete(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)

\method{chat_update}{default}(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)

\method{chat_delete}{default}(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)

\method{chat_update}{slackpost}(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)

\method{chat_delete}{slackpost}(
  post = NULL,
  channel = NULL,
  ts = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{character, Channel ID}

\item{...}{arguments to pass to API method.}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}

\item{post}{slackpost class object that contains meta information to identify a post.}

\item{ts}{character, timestamp}
}
\value{
A list with an additional class corresponding to \code{slack_method}.
}
\description{
Send, Edit or Delete a post from a channel based on the timestamp (ts).
}
\details{
Arguments that can be passed include 'text', 'blocks'
}
