% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{snippet_post}
\alias{snippet_post}
\alias{file_post}
\alias{file_delete}
\title{Interact with Files}
\usage{
snippet_post(content, channels, ..., token = Sys.getenv("SLACK_API_TOKEN"))

file_post(file, channels, ..., token = Sys.getenv("SLACK_API_TOKEN"))

file_delete(file, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{content}{character, content of snippet}

\item{channels}{character, Channel ID(s) to post file}

\item{...}{arguments to pass to API method.}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}

\item{file}{character, file ID}
}
\value{
A list with an additional class corresponding to \code{slack_method}.
}
\description{
Send, list, retrieve information or Delete a file.
}
\details{
For a full list of arguments to pass to file methods see \href{https://api.slack.com/methods}{here}
}
\examples{
\dontrun{
tf <- tempfile(fileext = '.r')
cat(
  utils::capture.output(utils::sessionInfo()),
  file = tf,
  sep = '\n'
)

#post the file
file_post(
  channels = "general",
  token = Sys.getenv('SLACK_API_TOKEN'),
  file = tf,
  filename = 'sessionInfo.R',
  filetype = 'r',
  title = 'R sessionInfo'
)

# delete the file
file_delete(file_last())
}
}
